NRSBtoDB-lambda-sam/NRSBtoDB/requirements.txt
requests
boto3
datetime
aws-psycopg2